Remove spam from generated firewall
The version number causes a lot of changes in a dry-run when oinkwall
version changes, remove it. Remove the whole line, who cares.

Remove vim hint about iptables syntax. I haven't used this for years,
and it's better to fix it in your own vim config if you want it. The
syntax highlighting instructions are also not included here.