Fix using custom chain in IPTablesRuleset
Custom chain is defined on the firewall object, not on the ruleset. So
it would never be possible to put rules into the custom chain. O_o

We used this to create a custom chain (without rules) where rules could
be dynamically inserted (like, think of some fail2ban like
construction).

When the ruleset is added to the firewall, the chain validation will
also be done.