Deduplicate results from TXT lookups
DNS TXT records like supported by this lib can also be used for other
purposes. Tooling might want to discover the network range a particular
IP address is in for example. Doing so requires smaller subnets to be
listed, e.g. 2001:db8:77:1::/120, which is part of 2001:db8:77::/48.
However, in a firewall, adding the /120 would be redundant if the /48 is
already in the list.

For example... this:

$ORIGIN collapse-example.knorrie.org.
_net      60 IN TXT "_net_1.collapse-example.knorrie.org"
          60 IN TXT "_net_2.collapse-example.knorrie.org"
          60 IN TXT "2001:db8:77::/48"
          60 IN TXT "203.0.113.16/28"
_net_1    60 IN TXT "_net_1a.collapse-example.knorrie.org"
          60 IN TXT "2001:db8:77:2::/120"
_net_1a   60 IN TXT "192.0.2.0/24"
_net_2    60 IN TXT "203.0.113.0/28"
          60 IN TXT "192.0.2.0/25"
          60 IN TXT "2001:db8:77:4::/120"

Will result in:

    import oinkwall
    oinkwall.firewall.parse_address_list(['_net.collapse-example.knorrie.org'])
    (['192.0.2.0/24', '203.0.113.0/27'], ['2001:db8:77::/48'])